requirements.txt
torch
transformers
faiss-cpu
pandas
numpy
scikit-learn
langchain-community
ollama
gradio
datasets
arabic-reshaper
python-bidi
